splitted page header in different comoponents parts to simplify header customizations
